+++ packages/code/src/Code.story.tsx
@@ -1,7 +1,7 @@
 /* eslint-disable no-useless-escape */
 import React from 'react';
 import { storiesOf } from '@storybook/react';
-import { select, boolean, text } from '@storybook/addon-knobs';
+import { select, boolean, text, number } from '@storybook/addon-knobs';
 import { css } from '@leafygreen-ui/emotion';
 import LeafyGreenProvider from '@leafygreen-ui/leafygreen-provider';
 import { Language } from './types';
@@ -27,7 +27,7 @@ export default class myClass {
 }
 
 function greeting(entity) {
-  return \`Hello, \${entity}!\`;
+  return \`Hello, \${entity}! Cras justo odio, dapibus ac facilisis in, egestas eget quam. Vestibulum id ligula porta felis euismod semper.\`;
 }
  
 console.log(greeting('World'));
@@ -38,10 +38,32 @@ storiesOf('Code', module)
   .add(
     'Multiline',
     () => {
-      const margin = 50;
+      // Knobs
+      const darkMode = boolean('darkMode', false);
+      const showLineNumbers = boolean('Show line numbers', false);
+      const lineNumberStart = number("First row's line number", 1);
+      const showWindowChrome = boolean('Show window chrome', false);
+      const showCustomActionButtons = boolean(
+        'Show custom action buttons',
+        false,
+      );
+      const copyable = boolean('Copyable', true);
+      const chromeTitle = text('Chrome label', 'directory/fileName.js');
+      const language = select(
+        'Language',
+        Object.values(Language),
+        Language.JavaScript,
+      );
+      const highlightLines = select(
+        'highlight lines',
+        ['none', 'single', 'multiple'],
+        'none',
+      );
+      const baseFontSize = select('Base Font Size', [14, 16], 14);
+      const codeSnippet = text('Code snippet', jsSnippet);
+
       const wrapperStyle = css`
-        margin: ${margin}px;
-        max-width: calc(100% - ${margin * 2}px);
+        width: 512px;
       `;
 
       const lineHighlightOptions = {
@@ -49,8 +71,6 @@ storiesOf('Code', module)
         single: [1],
         multiple: [[2, 4], 6],
       } as const;
-
-      const darkMode = boolean('darkMode', false);
 
       const actionData = [
         <IconButton
@@ -74,35 +94,21 @@ storiesOf('Code', module)
       ];
 
       return (
-        <LeafyGreenProvider>
+        <LeafyGreenProvider baseFontSize={baseFontSize}>
           <div className={wrapperStyle}>
             <Code
-              showLineNumbers={boolean('Show line numbers', false)}
-              showWindowChrome={boolean('Show window chrome', false)}
-              showCustomActionButtons={boolean(
-                'Show custom action buttons',
-                false,
-              )}
+              showLineNumbers={showLineNumbers}
+              showWindowChrome={showWindowChrome}
+              lineNumberStart={lineNumberStart}
+              showCustomActionButtons={showCustomActionButtons}
               customActionButtons={actionData}
-              copyable={boolean('Copyable', true)}
-              chromeTitle={text('Chrome label', 'directory/fileName.js')}
+              copyable={copyable}
+              chromeTitle={chromeTitle}
               darkMode={darkMode}
-              language={select(
-                'Language',
-                Object.values(Language),
-                Language.JavaScript,
-              )}
-              highlightLines={
-                lineHighlightOptions[
-                  select(
-                    'highlight lines',
-                    ['none', 'single', 'multiple'],
-                    'none',
-                  )
-                ]
-              }
+              language={language}
+              highlightLines={lineHighlightOptions[highlightLines]}
             >
-              {text('Code snippet', jsSnippet)}
+              {codeSnippet}
             </Code>
           </div>
         </LeafyGreenProvider>
@@ -140,11 +146,20 @@ storiesOf('Code', module)
 
     return (
       <LeafyGreenProvider>
-        <LanguageSwitcherExample
-          darkMode={darkMode}
-          showCustomActionButtons={boolean('Show custom action buttons', false)}
-          customActionButtons={actionData}
-        />
+        <div
+          className={css`
+            width: 512px;
+          `}
+        >
+          <LanguageSwitcherExample
+            darkMode={darkMode}
+            showCustomActionButtons={boolean(
+              'Show custom action buttons',
+              false,
+            )}
+            customActionButtons={actionData}
+          />
+        </div>
       </LeafyGreenProvider>
     );
   });